Refactor: MAT-350 이어쓰기 설정 모달에서 View만 남기기 위해서 ws 의존성 제거
